Preparation for job queueing and abstraction of compute sessions
* refs:
  - lablup/backend.ai-manager#110
  - lablup/backend.ai-manager#111

Logic:

All kernel creation requests will by default be added to a pending
request table, in the form of SessionRequest.

The job scheduler will fetch them and choose the set of next-executed
SessionRequests and the agents responsible for them.

TODO: It would be nice to have the followings implemented as intrinsic
      APIs in the manager and/or the common.
  - Matching a resource request against the resource specification
    reported by agents
  - (De-)serialization of these requests/spec objects for database
    storage and networking.